Convert arguments to unicode. Fix for issue #2
Using a search term with an accent didn't find anything, and giving a
note name with an accentuated character in it resulted in an exception
going up unhandled.

To fix both of these issues, one change has been made: arguments after
the action name are now converted to unicode objects automatically,
before being sent to actions.

We now expect the arguments from the command line to be encoded in
UTF-8.