Option and OptionGroup abstraction.
Abstract optparse.Option and optparse.OptionGroup away from ActionPlugin
subclasses.

I've finally rolled this one out. Touches almost all files in the code.
This is the second of three steps to get to option libraries.
ActionPlugin subclasses can now use two methods from their ancestor to
add options and to group them.

This abstraction was needed because of optparse.OptionGroup's
constructor's need of a lousy OptionParser in the arguments.
tomtom.plugins.OptionGroup is the intermediary class. It simply keeps a
hold of values until we can have a parser to create a real
optparse.OptionGroup object.

Option groups should be added before options are added to them. Both
methods (ActionPlugin.add_option and ActionPlugin.add_group) accept the
same arguments as optparse.Option and optparse.OptionGroup respectively
-- well, in fact they simply pass arguments on. ActionPlugin.add_option
intercepts the keyword argument "group" to be able to determine which
option group the new option should be attached to. If this keyword
argument is not present, the option is attached to the "None" group so
that it does not belong to any group at all.

Filtering options were grouped for the eventuality of being aggregated
in the next commit.

One important change is that a Tomtom object is automatically created as
the attribute "tomboy_interface" in ActionPlugin subclasses right before
their perform_action is called. This made two acceptance tests fail
because of the DBus connection it creates. Those two particular cases
(display with no arguments, search with no arguments) were not too
problematic because perform_action has to be called before we know if
things are right or not. It should not have a big impact on performance,
plus both of those cases output an error message.

Coverage was a little bit improved: unit tests for cli.main() were
created. A return code was added in the process to the error shown when
tomtom is called without any argument.