refactor: move action code away from core.Tomtom
core.Tomtom was built in a weird fasion. It contained code that really
belonged to the action plugins. Now that the action plugins have a clean
design, move this code in their appropriate place.

Tomtom now contains two "screen" methods that do nothing but return the
result of calling the method with the same name in the
Tomboy_communicator object. The right way to fix things would be to move
everything from Tomboy_communicator to Tomtom, effectively removing the
need for Tomboy_communicator. However, this is another big change and it
will be done in the next commit to avoid overcrowding this one.

- Move listing method to ListAction and simplify method calls.

- Move get_display_for_notes to DisplayAction and change the name to
  something more evocative (format_display_for_notes)

- Move search_for_text to SearchAction

- Adjust all tests accordingly

- Create two "screen" methods in Tomtom to prepare for the next commit.

- Refactor verify_note_listing to use a TomboyNote mock object instead
  of instantiating one. Control during the test is better this way.