Add --gnote to suuport connection to Gnote
Gnote aims to be a 1:1 clone of Tomboy. Simply changing the strings used
to connect to DBus, it is possible to interact out of the box with
Gnote.

Add support for Gnote by providing a way to switch the application we
connect to. Also change output to print Gnote instead of Tomboy when we
are connecting to Gnote. This way, it is clearer what Tomtom is doing.