Fix 2 Oopses
Seems like my commit from yesterday was made without running the tests
first. The expected help texts were not adjusted to the change.

Also, commit 4f2e58ba2c7748b562e91a3306c7d7954e5d7d29 moved the
Tomtom object instantiation inside dispatch (and, thus, outside of
perform_action) but did not adjust the tests for this. As a result,
ConnectionError was going unhandled.

When a request is made to dbus when Tomboy or Gnote are not running,
they get started automatically. So the error message about "Is the
application started?" is incorrect. It should talk more about the
DISPLAY environment variable.

Fix both these problems and change the connection error message.