Change --gnote into --application
After a discussion on the Gnote mailing list, it was agreed that
autodetection should be added via dbus. In a context where the final
choice for choosing to which application we connect to is autodetection,
having only an option that can force using gnote is not sufficient.

Change the option --gnote into --application and make it into a textual
choice instead of a binary option. This way the user can specify
--application=Tomboy or --application=Gnote.

By using the "choices" argument to optparse.Option, make sure the value
space for that option is very limited (in this case 3 possibilities:
None, "Tomboy", "Gnote"). Also, if the user uses something other than
the 3 mentioned values, present an error message with the possible
choices.