Application autodetection
In order to support more gracefully Tomboy and Gnote, add a mechanism to
simplify using tomtom. Since most users will not have both Tomboy
__and__ Gnote installed on their computers, it is a sensible default
that tomtom tries to determine automatically which one to use.

To implement autodetection, calls to get_object for both applications on
the dbus proxy object give us information about which application is
available or not. get_object automatically starts the application that
is requested. This has the weird effect of starting up both Tomboy
and Gnote before outputting an error message if both are present.

Autodetection succeeds if only one application is detected.

Autodetection is preceded by the command line option --application so
that the user can keep control over what is happening.

+ Fixed the mock_out_dbus function in test_acceptance: the way the
  SessionBus and Interface classes were mocked out, we were loosing the
  original classes in test_list_using_gnote by invoking mock_out_dbus a
  second time. Fix this by using mox to stub out the classes.