Configuration file: choose app. -- closes #19
One part was missing for the Gnote support to be complete: the
possibility for users to use a configuration file to specify which one
of Tomboy and Gnote is the preferred application, instead of always
having to use the "--application" command line argument.

Add retrieval of a configuration file. Keep only the values that are
known to the program in the "tomtom" section. Check if a value was given
to the "application" configuration option.