testing: missing test for Tomtom's constructor
A unit test was missing for Tomtom's constructor: when an application
name is specified but the application does not respond, an exception
should be raised.

Add the missing test. At the same time, remove clunky testing code for
mocking out the dbus SessionBus and Interface classes. Use mox to mock
out the classes correctly.

+ Add a commented line in nose.cfg for running acceptance tests only.