New action: delete -- closes #14
Tomtom didn't know yet how to delete notes.

Add this functionality with a new action called "delete".

The delete action has a --dry-run option which prevents it from making
any modifications to notes. Instead of actually deleting the notes, it
will print a message on the standard output to list which notes would be
deleted without the option.

This action also modifies the default behaviour for templates inclusion
in the note filtering mechanism. By default, templates are included
implicitly (rather than explicitly, as in other actions). This
facilitates deletion of a whole book by specifying "-b the_book". If
templates must not be deleted, the --spare-templates option must be
supplied on the command line.

Thus, deleting notes can be done by invoking tomtom in the following
manner:

    tomtom delete -b the_book "additional note" trash_note

A big change was necessary to remove an odd specification in the note
filtering mechanism. Specifying note names prevented from specifying
tags at the same time. This limitation (a previous design choice,
though) was not hit by any actions yet. The code responsible for note
retrieving and filtering was simplified. Some functions were removed.

With the old mechanism, specifying note names helped in cutting out the
number of processed notes by making a series of call to FindNote instead
of ListAllNotes. This is no longer the case, since specifying note names
makes filtering include the notes with those names instead of
restricting processing on this set of notes.

+ In acceptance tests: Changed mock_out_get_notes_by_names for
  mock_out_listing, since get_notes_by_names was removed.
+ Added the use of "metavar" arguments for tag and book filtering so
  that the help message is less confusing. Each use of the option take
  one tag and one book, respectively. Displayed variable names are now
  TAG instead of TAGS and BOOK instead of BOOKS.
+ Changed the filtering options' order to place --with-templates at the
  end.
+ Changed the other actions' usage strings to make them simpler to read.