delete without arguments erases all notes
Giving no filtering argument and no note name to the "delete" action
selects all notes for deletion.

To reproduce, use "tomtom delete --dry-run". This will output the
complete list of notes.

This is a very easy way for users to make a big mistake. Change delete's
perform_action method to do nothing when no argument is given, and
output a message to guide users.