teach --all-notes to action delete -- closes #22
With the previous commit, deleting all notes at once was rendered
impossible without listing all tags and note names of untagged notes.

Add an option to bring back this feature. The option is --all-notes. By
making it explicit, it helps to avoid the deletion of all notes by
simply calling the delete action with no argument.