Version bump. Release of 0.3
This marks the release of Tomtom 0.3 This release has seen lots and lots
of code rework to make it better / more logical. Also, unit test
coverage was boosted to 100%.

Output is now forced to being encoded in UTF-8. This makes it possible
to reuse the output by piping it to another command or by redirecting it
to a file. I didn't have any feedback on "forcing UTF-8" being a bad
thing yet, so with the lack of testers, the solution (/workaroud) was
not complexified or made more intelligent.

Important new features were also added. Among them the following stand
out:

    Actions are now setuptools plugins
    Support for working with Gnote, since the dbus API is the same.
    New actions for general information: help, version
    New action for note deletion: delete

An entry point for actions is used to retrieve all the available
actions. This removes the ugly code using __import__ and makes it even
easier to develop new actions without requiring them to be in Tomtom's
directory. Plus, setuptools is now able to compress the egg file upon
installation, saving some precious disk space.

Tomtom automatically detects which application to connect to, if only
one of Tomboy and Gnote is installed. A configuration file can also be
used to specify the default application in case of a conflict (both
applications installed), and a command line option can also be used to
force Tomtom to use an application.

The "help" action is a "pseudo-action", since there is no entry point
defined for it. It is simply an alias from 'tomtom help action' to
'tomtom action -h'. The name "help" must thus be avoided as an action
name because it will never be reached.

The "version" action currently only displays information on Tomtom's and
Tomboy's or Gnote's version tags. It is a useful means to know which
application Tomtom connects to.

The "delete" action inverses the default behaviour about note inclusion.
When called with no argument, nothing is done. There is no
--with-templates option since templates are included by default (being
the opposite of what the other options do). An option named
--spare-templates can be used to exclude templates from deletion (so in
fact, keep them).