Better ctrl-c handling
In a particular case, it was still possible for tomtom to spit a traceback
for KeyboardInterrupt. This particular case is when ctrl-c is hit while
performing imports.

Also, exiting when user asks for premature termination should not
correspond to a successful execution. Tomtom currently models it as
such.

Add a try block around imports and constant definitions that catches
KeyboardInterrupt and make it exit with an error code of 1.

Also, change the "pass" on KeyboardInterrupt handling around main() to
make it exit with an error code of 1.

One situation still lacks ctrl-c handling: hitting it very fast at
startup time, while the distutils-generated wrapper around
"exception_wrapped_main" is doing its work. This situation is
unfortunately out of our control.

One more, minor change:
    In exception_wrapped_main, no reference to the CommandLine instance
    is kept anymore, since the object can be thrashed when main exits.