incorrect import renders application unusable
oh, silly me. When I did the cleanup for packaging, I forgot to change
the import directive in cli.py. I probably didn't catch this because I
had some modifications on cli.py in the work dir. so, my pre-commit hook
that runs tests is not infallible.

This change fixes the last import directive for TOMTOM_VERSION.