man pages + auto version tag
Add documentation for the main application. Also add a build mechanism
for the documentation files to transform them from Markdown to groff
format.

Move the version string again to a file that gets generated. For source
archives that are generated with the "git-archive" command,
gitattributes help in extracting the necessary values. When the source
code is inside its git repository, git is used to dynamically generate
the appropriate values.
This modification implies that the version tag inside TOMTOM_VERSION
will not need to be bumped manually (and possibly forgotten) with each
release from now on, since it will be generated automatically.

Adjust the LICENSE and README files to mention the different license
under which the "format-subst.pl" script is.
Also add a description of the program at the beginning of the README and
change its extension from ".markdown" to ".md".