Project renamed Scout
Due to forseable trademark issues, the project is renamed to Scout.

Move files and modify their contents so that references to "Tomtom" are
changed to "Scout".

Also: Added *.1.gz in the 'clean' target of the doc/Makefile so that
compiled manpages are removed and changed old reference to
"README.mardown" in doc/scout.1.md

And in the readme:

Update the README.md file to talk about the new mailing list.

Make some adjustments to better specify the context in which commands
should be run. Fix some typos and use better formulation.

Add a note in the Contributing section that asks people to send in
patches with "Signed-off-by:" lines in them.