Remove duplicate license notices
The LICENSE file already has what we want for indicating a license.
There is no need to copy its entirety in every code file. Remove this
crud.

The LICENSE file said that the format-subst.pl file was under GPLv2,
which is not right. It is under LGPLv2, as indicated in the file.
Correct the statement, and also append the full LGPLv2 text inside the
LICENSE file.

Remove useless docstrings inside __init__.py files. Keep only the
docstring inside the __init__.py file that is in the topmost source
directory so that doing help(scout) shows a little something. Also
change this docstring's text into a more general text about the
application.

The huge comment at the top of all action modules is not in the
appropriate spot and duplication is tough/boring to maintain. Remove
the whole comment and create a new scout-plugins.1 man page that
describes how to write action plugins.

The date inside the scout.1 man page is hardcoded, make it reflect the
export date by using the %SCOUT_DATE% magic token.