Janitor work part 2
The code cleanup continues. I'm trying to get closer to the
recommendations from PEP 8 and to get generally less clutter inside the
code.

This patch works mainly on core.py and actions/list.py but also touches
both test files.

Make docstrings say only relevant and to-the-point things.

Remove some comments that say what the code does (there are probably
some left): comments should explain why the code follows some weird
paths, not what it is actually doing since it is evident from reading
the code.

Use ''.join([]) instead of "" + "" whenever possible and factor out
relevant splitting chars inside the the leading string (e.g.
'\n'.join() instead of ''.join("...\n","...") ).

Rename class TomboyNote to Note to make it more generic since we don't
only support Tomboy anymore.

Change Note.listing() into Note.__repr__() so that a note can be printed
out directly. This represents better the purpose of the function and
makes the code that prints notes seem easier to understand.

Change os.linesep to "\n" since it is a shorter representation _and_ it
is already the portable representation of newlines (it is already
converted to the right thing on systems that use multiple characters to
represent new lines).

Signed-off-by: Gabriel Filion <lelutin@gmail.com>