Make setup.py develop use a more static version
Since the version tag was automated, when developing and using "setup.py
develop" is annoying: the version tag changes after each commit so the
installed link is invalid after creating a commit.

Trucate the version tag to remove the number of commits since last tag
and the current commit's hash, only when the setup.py script is called
with the "develop" command, and replace the missing part with -dev.

Signed-off-by: Gabriel Filion <lelutin@gmail.com>