Janitor work last part.
This marks the last one of a series of big style cleanup commits.

Work on setup.py, cli.py, core.py, plugins.py and all actions.

Also, I would've liked to split this in another commit but I messed up
my work dir with a git-stash command (and I don't want to resplit
everything):
Change the style of Ctrl-C handling to a more elegant one. Also remove
"exception_wrapped_main" in favor of cli.main(). Having to wrap
everything inside of an ugly-named function is gross. I lose a little
bit testability with this but I think having code that makes sense and
that is easier to read is more important.

Signed-off-by: Gabriel Filion <lelutin@gmail.com>