Rename test files and split up function tests.
This has bugged me for some time. I named the test files
"test_acceptance" and "test_application" since I was following a new
method. But the two are actually functional and unit tests,
respectively. Let's call an apple an apple and a bee a bee.

Refactor the unit tests file to put PluginsTest right after CoreTests so
that it is grouped better with main application functionality tests.

Also, refactor the functional tests so that they are split up in logical
sections, as is the case for unit tests. Adjust test names to show the
section in which they are.
And finally, the "standard" for test docstrings is changed to include a
letter in the beginning that tells what kind of test it is. "F" for
functional tests and "U" for unit tests.

Signed-off-by: Gabriel Filion <lelutin@gmail.com>