Add -T option to filter for untagged notes
Currently, there is no mechanism to list notes that have no tags
exclusively. This is related to issue #28.

Make filter_notes a little more generic by using a list of unnamed
lambdas instead of enumerating calls to the named lambdas. This looks
nicer and makes it possible to extend filters by adding lambdas to the
list.

Use the new possibility to add a filter when None is contained in the
required tag list. Also add a -T option that adds None to the list of
required tags.

Signed-off-by: Gabriel Filion <lelutin@gmail.com>