Force DISPLAY through argument or config. closes #27
The aim of this project is to be used on the command line. When we are
not in an X session, for example via SSH, the DISPLAY environment
variable is not set and we need to explicitly export it.

Add a new global argument '--display' and a new configuration option
that lets users force the value for this environment variable.

Signed-off-by: Gabriel Filion <lelutin@gmail.com>