New -B filter gets unbooked notes -- closes #28
Add a functional test for the -T option that was added in
4e95126d2.

Add a -B option to the command line and make it select notes that are
not part of any books. (In technical terms, notes that have no tag
starting with "system:notebook:")
We need to have another "special" value to represent "no-book". None is
great for "no-tag" since it can be read as such and it is not a string.
But for books, it was not as straight forward. So, I added a class
NoteBook in which the name of the book is separated from the prefix.
This way we can represent "no-book" by a NoteBook instance with an empty
name.

A new method was added to Note to retrieve all the "book"-tags. We can
test this list for truthfulness in order to decide whether a note is
booked or not.

Also, refactor functional tests for filtering. Since they all test
against the 'list' action, they all do pretty much the same thing.

Finally, full_list_of_notes with mock objects (default behaviour) was
acting strange. It seems the caching of the mock objects is actually a
bad thing because the Mox factory gets recreated before each test. This
means we can keep Note mocks that were attached to a very different
factory.
To make it behave correctly, remove caching for mocks.

Signed-off-by: Gabriel Filion <lelutin@gmail.com>