Fix version tag generation
Last refactor on setup.py moved the line that imports the version #
above all code. This broke version.py generation.

Move the code that creates version.py above the import to ensure that
the file exists before we use it. Also wait() on the subprocess to
ensure that creation was completed.
With a comment above the generation code, this mistake should not be
happening again.

Signed-off-by: Gabriel Filion <lelutin@gmail.com>