Add URLs to Tomboy's and Gnote's DBus code
In the past, I removed the URL to the an article talking about using
Tomboy's DBus interface, but it was pretty useful.

Add it back in core.py and also add links to files, in the Tomboy and the
Gnote git repositories, that define their DBus interface. This should
prove even more useful than the article.

Signed-off-by: Gabriel Filion <lelutin@gmail.com>