Small adjustments to the 'delete' action
All actions mention both -h and --help in their help message and
'delete' shouldn't be an exception. Change the usage text in order to
mention the long help option.

The error message shown when no argument is given is currently sent to
stdout. Change this to go to stderr, like an error message should.

Signed-off-by: Gabriel Filion <lelutin@gmail.com>