Teach 'edit' to remove tags
Now that the 'edit' action knows how to add tags to notes, we'd like to
be able to remove them.

The code path was already laid out by the tag addition feature so
plugging in tag removal was simple, but removing tags has more pitfalls
than adding them, so we need to make sure we're doing something
legitimate.

One can add tags and remove others in the same operation but to avoid
confusion where someone adds and removes the same tag in the same
operation and nothing has apparently changed, we bail out when such an
operation is requested.

Signed-off-by: Gabriel Filion <lelutin@gmail.com>