edit: fix file open/close operations
since we're now using python3, the temp file now gets created using
binary mode by default and the write operation fails since we're giving
it a string content to write. to fix this we have to reopen the file
without the "b" flag.

additionally the file descriptor was not being closed after reading was
completed before setting new note content.