avoid useless encodings to utf8 (Closes:#34)
This is a followup for the port to python3. In py3, all string objects
are utf8 by default so we don't need to force encoding to it.

All of the actions that were modified were displaying some weirdly
formatted strings because of this instead of cleanly showing the
contents that we wanted. It had the effect of showing escaped special
chars like \n instead of actually displaying them.