Move to pyproject.toml
Using setup.py directly is deprecated. We can now concentrate everything
into that new file, so both requirements*.txt files are no longer
useful.

I found out the trick of using setuptools_scm by looking at the
pyproject.toml file of pytest so I migrated to using that file instead
and that same trick for dynamic version string. It does essentially the
same thing as what was being done previously, with a somewhat different
result. This method has the advantage of not being tied in to git, so
it's better.

I've removed the use of format-subst.pl since we don't need it anymore.